Fix defects in the repaso exercise solutions
The project did not compile: 365 does not fit in a byte (2.1).

Some solutions printed a wrong answer. The palindrome check ignored
accents and failed on the example of its own statement (4.10), the
decimal validator rejected signed numbers and accepted a lone
separator (4.7), and 8.5 staged an ArithmeticException that could
never be reached.

Others did not enforce the rule the exercise exists to teach. MyDate
validated only setDay, so setMonth and setYear could build 31/2 or
29/2 of a common year (8.3). A bank account accepted negative amounts,
so a withdrawal increased the balance and a transfer drained the
destination (7.7). Measurement had no equals nor hashCode, which left
every Set of section 5 holding objects that a student reads as equal.

Person now receives the DNI in every constructor except the empty one,
which generates a random one, following the amended statement. Its
main asks again for a weight or a height that is not positive, since
the exception that would guard computeBMI belongs to section 8.

The rest are the edge cases the introduction asks for: empty rows in
3.6 and 5.re06, and trailing empty fields in the CSV of 8.2. Also the
menu that 7.3 requires and was missing, and the append of 6.3, which
read and rewrote the whole file.

+++ src/es/upm/dit/repaso/ej201/Main.java
@@ -8,8 +8,9 @@ package es.upm.dit.repaso.ej201;
 public class Main {
 
 	public static void main(String[] args) {
-		// Usar tipos pequeños para economizar memoria
-		byte diasPorAnio = 365;
+		// El tipo más pequeño que admite cada valor: 365 no cabe en un byte
+		// (rango -128..127), los demás sí
+		short diasPorAnio = 365;
 		byte horasPorDia = 24;
 		byte minutosPorHora = 60;
 		byte segundosPorMinuto = 60;
